refactor: Separate tests from benchmarks
- test_optimizations.py: Correctness tests (runs in CI)
- benchmark_optimizations.py: Performance benchmarks (deletable)

Steered and verified with @nikblanchet
